add the multer function to the jobcontroller and update the roles

+++ src/jobs/jobs.service.ts
@@ -1,10 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateJobDto } from './dto/create-job.dto';
 import { UpdateJobDto } from './dto/update-job.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { JobEntity } from './entities/job.entity';
 import { CompanyEntity } from 'src/company-management/entities/company-management.entity';
 import { Repository } from 'typeorm';
+import { UserEntity, UserRole } from 'src/user/entities/user.entity';
 
 @Injectable()
 export class JobsService {
@@ -28,18 +29,24 @@ export class JobsService {
   }
 
 
-  async updateJob(id: number, updateDto: Partial<CreateJobDto>): Promise<JobEntity> {
+  async updateJob(id: number, updateDto: Partial<CreateJobDto>,user:any): Promise<JobEntity> {
     const job = await this.jobRepository.findOne({ where: { id } });
     if (!job) throw new NotFoundException('Job not found');
+
+    if((job.company.user.id!==user.id)||(user.role!==UserRole.ADMIN)){
+      throw new ForbiddenException('You cannot edit this job');
+    }
 
     Object.assign(job, updateDto);
     return await this.jobRepository.save(job);
   }
 
-  async deleteJob(id: number): Promise<{ message: string }> {
+  async deleteJob(id: number,user:any): Promise<{ message: string }> {
     const job = await this.jobRepository.findOne({ where: { id } });
     if (!job) throw new NotFoundException('Job not found');
-
+    if((job.company.user.id!==user.id)||(user.role!==UserRole.ADMIN)){
+      throw new ForbiddenException('You cannot delete this job');
+    }
     await this.jobRepository.remove(job);
     return { message: 'Job deleted successfully' };
   }
